update MapOverlayView and move it to the MapFeature
* fix animation by using .onChange and internal @State properties
